requirements.txt
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.3.2
pytz==2021.3
sqlparse==0.4.2
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-multi-form-view==2.0.1
Pillow==8.3.2
python-decouple==3.8
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
asgiref==3.4.1
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-multi-form-view==2.0.1
Pillow==8.3.2
python-decouple==3.8
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.6.0
